+++ src/cpu/x64/jit_avx512_core_x8s8s32x_conv_kernel.cpp
@@ -1733,7 +1733,7 @@ status_t jit_avx512_core_x8s8s32x_fwd_kernel::init_conf(jit_conv_conf_t &jcp,
     const auto &src_scales = attr.scales_.get(DNNL_ARG_SRC);
     const auto &wei_scales = attr.scales_.get(DNNL_ARG_WEIGHTS);
     const auto &dst_scales = attr.scales_.get(DNNL_ARG_DST);
-    const int wei_mask_per_oc = 1 << with_groups;
+    const int wei_mask_per_oc = 1 << (int)with_groups;
     jcp.is_oc_scale = wei_scales.mask_ == wei_mask_per_oc;
     jcp.dst_scale = !dst_scales.has_default_values();
