+++ src/cpu/x64/jit_avx512_core_x8s8s32x_deconvolution.cpp
@@ -282,7 +282,7 @@ status_t _jit_avx512_core_x8s8s32x_deconv_fwd_kernel::init_conf(
     const auto &src_scales = attr.scales_.get(DNNL_ARG_SRC);
     const auto &wei_scales = attr.scales_.get(DNNL_ARG_WEIGHTS);
     const auto &dst_scales = attr.scales_.get(DNNL_ARG_DST);
-    const int wei_scales_per_oc = 1 << with_groups;
+    const int wei_scales_per_oc = 1 << (int)with_groups;
     jcp.is_oc_scale = wei_scales.mask_ == wei_scales_per_oc;
     jcp.dst_scale = !dst_scales.has_default_values();
