+++ src/cpu/x64/jit_brgemm_conv_utils.cpp
@@ -2079,7 +2079,7 @@ status_t init_conf(jit_brgemm_conv_conf_t &jcp, cpu_isa_t isa,
     const auto &dst_scales = attr.scales_.get(DNNL_ARG_DST);
     jcp.with_scales = !src_scales.has_default_values()
             || !wei_scales.has_default_values();
-    const int wei_mask_per_oc = 1 << with_groups;
+    const int wei_mask_per_oc = 1 << (int)with_groups;
     jcp.is_oc_scale = wei_scales.mask_ == wei_mask_per_oc;
 
     // only common and per-oc-channel scales are supported
@@ -2319,7 +2319,7 @@ status_t init_1x1_conf(jit_brgemm_conv_conf_t &jcp, cpu_isa_t isa,
     const auto &dst_scales = attr.scales_.get(DNNL_ARG_DST);
     jcp.with_scales = !src_scales.has_default_values()
             || !wei_scales.has_default_values();
-    const int wei_mask_per_oc = 1 << with_groups;
+    const int wei_mask_per_oc = 1 << (int)with_groups;
     jcp.is_oc_scale = wei_scales.mask_ == wei_mask_per_oc;
 
     // only common and per-oc-channel scales are supported
